Factorize and fix read/write FCB functions.
Use only the first 3 bytes of the random position if the block size is
more than 64 bytes, as documented in Ralf Brown's interrupt list, and
only update random position if actually doing random reads.

This fixes some corner cases in FCB functions, and makes the behaviour
the same as MS-DOS from version 1.0 to 5.0.

This correctly closes #20.